# optimize cohorts yet again