```
| Before [ac319cc2]| After [e87be3e4] |   Ratio | Benchmark (Parameter)                       |
|------------------|------------------|---------|---------------------------------------------|
| 28.2±0.4ms       | 23.6±0.1ms       |    0.84 | cohorts.OISST.time_graph_construct          |
| 6.24±0.02ms      | 4.97±0.03ms      |    0.8  | cohorts.ERA5Google.time_graph_construct     |
| 294±0.8ms        | 219±4ms          |    0.75 | cohorts.ERA5Resampling.time_graph_construct |
| 42.2±0.7ms       | 29.9±0.1ms       |    0.71 | cohorts.ERA5DayOfYear.time_graph_construct  |
```